Capitalization and other typos/rewording.
How long are lines limited to? One word I added to a question made it 78(?) characters long.
Don't understand the sentence "While that set of a bit verbose, it's easy to remember
since you're doing one thing at a time." so didn't try to correct it. Please check my correction to the question above it.